simulation/gz: fix float to double promotions
These are -Wdouble-promotion errors under PX4's -Werror set, hit wherever
PX4's float values meet the gz::math / gz::msgs double-precision API.

MovingPlatformController: max_accel, accel_xy and scaling were float in
an otherwise all-double block (the surrounding gains, masses and forces
are double, and _platform_mass is double). Making them double matches
the surrounding code and removes the promotion at the Vector3d call.

GZMixingInterfaceServo: get_servo_angle_min/max return float and are
stored in std::vector<double>. GZGimbal: a float setpoint is passed to
gz::msgs::Double::set_data(). Both widen deliberately, so make the
conversion explicit.

No behavioural change.

+++ src/modules/simulation/gz_bridge/GZGimbal.cpp
@@ -231,7 +231,7 @@ void GZGimbal::publishJointCommand(gz::transport::Node::Publisher &publisher, co
 	float new_stp = computeJointSetpoint(att_stp, rate_stp, last_stp, dt);
 	new_stp = math::constrain(new_stp, min_stp, max_stp);
 	last_stp = new_stp;
-	msg.set_data(new_stp);
+	msg.set_data(static_cast<double>(new_stp));
 
 	publisher.Publish(msg);
 }
